pyproject.toml
[tool.poetry]
name = "clipper"
version = "0.4.0"
description = ""
authors = ["MaxMixAlex <MaxMixAlex@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyperclip = "^1.8.2"
rich = "^13.6.0"
textual-dev = "^1.2.1"
requests = "^2.31.0"
watchdog = "^4.0.1"
show-in-file-manager = "^1.1.5"

[tool.poetry.scripts]
clipper = "clipper.__main__:main"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
